WSIO-708 extract BlogArticleHeader interface
