+++ applications/blogs/backend/src/main/java/pl/ds/kyanite/blogs/components/models/BlogArticleHeaderComponent.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2023 Dynamic Solutions
+ * Copyright (C) 2024 Dynamic Solutions
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,79 +16,13 @@
 
 package pl.ds.kyanite.blogs.components.models;
 
-import java.util.Objects;
-import javax.annotation.PostConstruct;
-import javax.inject.Inject;
-import lombok.Getter;
-import org.apache.sling.api.resource.Resource;
-import org.apache.sling.api.resource.ResourceResolver;
-import org.apache.sling.models.annotations.DefaultInjectionStrategy;
-import org.apache.sling.models.annotations.Model;
-import org.apache.sling.models.annotations.injectorspecific.OSGiService;
-import org.apache.sling.models.annotations.injectorspecific.SlingObject;
-import org.apache.sling.models.factory.ModelFactory;
-import pl.ds.kyanite.blogs.components.exceptions.AuthorInfoResolvingException;
-import pl.ds.kyanite.blogs.components.services.AuthorInfoResolverService;
-import pl.ds.kyanite.blogs.components.utils.ResourceUtil;
 import pl.ds.kyanite.common.components.models.layouts.TemplateHandler;
 
-@Model(
-    adaptables = {Resource.class},
-    defaultInjectionStrategy = DefaultInjectionStrategy.OPTIONAL
-)
-public class BlogArticleHeaderComponent implements TemplateHandler {
+public interface BlogArticleHeaderComponent extends TemplateHandler {
 
-  private final Resource resource;
+  BlogArticleHeaderModel getBlogArticleHeaderModel();
 
-  private final ResourceResolver resourceResolver;
+  AuthorInfoModel getAuthorInfoModel();
 
-  private final ModelFactory modelFactory;
-
-  @Getter
-  private BlogArticleHeaderModel blogArticleHeaderModel;
-
-  @Getter
-  private AuthorInfoModel authorInfoModel;
-
-  @Getter
-  private String authorInfoErrMessage;
-
-  private final AuthorInfoResolverService authorInfoResolver;
-
-  @Inject
-  public BlogArticleHeaderComponent(
-      @SlingObject Resource resource,
-      @SlingObject ResourceResolver resourceResolver,
-      @OSGiService ModelFactory modelFactory,
-      @OSGiService AuthorInfoResolverService authorInfoResolver
-  ) {
-    this.resource           = resource;
-    this.resourceResolver   = resourceResolver;
-    this.modelFactory       = modelFactory;
-    this.authorInfoResolver = authorInfoResolver;
-  }
-
-  @PostConstruct
-  private void init() {
-    resolveHeaderModel();
-    resolveAuthorInfo();
-  }
-
-  private void resolveHeaderModel() {
-    final Resource currentPage = ResourceUtil.getParentPageResource(resource, resourceResolver);
-    if (Objects.nonNull(currentPage)) {
-      final Resource pageContent = ResourceUtil.getContentNode(currentPage);
-      blogArticleHeaderModel = modelFactory.createModel(pageContent, BlogArticleHeaderModel.class);
-    }
-  }
-
-  private void resolveAuthorInfo() {
-    try {
-      authorInfoModel = authorInfoResolver.retrieveAuthorInfo(resource, resourceResolver);
-    } catch (AuthorInfoResolvingException e) {
-      authorInfoErrMessage = e.getMessage();
-    }
-  }
-
+  String getAuthorInfoErrMessage();
 }
-
