+++ applications/blogs/backend/src/main/java/pl/ds/kyanite/blogs/components/models/BlogListComponent.java
@@ -40,6 +40,7 @@ import org.apache.sling.models.annotations.injectorspecific.SlingObject;
 import org.apache.sling.models.annotations.injectorspecific.ValueMapValue;
 import org.apache.sling.models.factory.ModelFactory;
 import org.jetbrains.annotations.Nullable;
+import pl.ds.kyanite.blogs.components.models.impl.BlogArticleHeaderModelImpl;
 import pl.ds.kyanite.blogs.components.services.AuthorInfoResolverService;
 import pl.ds.kyanite.blogs.components.services.BlogArticleService;
 import pl.ds.kyanite.common.components.utils.LinkUtil;
@@ -129,7 +130,7 @@ public class BlogListComponent {
             BlogArticle.builder()
                 .link(LinkUtil.handleLink(
                     StringUtils.substringBefore(res.getPath(), JCR_CONTENT), resourceResolver))
-                .blogArticleHeader(modelFactory.createModel(res, BlogArticleHeaderModel.class))
+                .blogArticleHeader(modelFactory.createModel(res, BlogArticleHeaderModelImpl.class))
                 .authorInfo(retrieveAuthorInfo(res))
                 .build()
         )
